+++ mobile/cmp/tab/impl/TabContainer.js
@@ -7,29 +7,44 @@
 
 import {Component} from 'react';
 import {HoistComponent, elemFactory} from '@xh/hoist/core';
-import {tab as onsenTab, tabbar} from '@xh/hoist/kit/onsen';
-import {tab} from './impl/Tab';
-import {TabContainerModel} from './TabContainerModel';
 import {div} from '@xh/hoist/cmp/layout';
+import {tab as onsenTab, tabbar} from '@xh/hoist/kit/onsen';
+import {TabContainerModel} from '@xh/hoist/cmp/tab';
+import {throwIf} from '@xh/hoist/utils/js';
+import {tab} from './Tab';
 
+/**
+ * Mobile Implementation of TabContainer.
+ *
+ * @private
+ */
 @HoistComponent
 export class TabContainer extends Component {
 
     static modelClass = TabContainerModel;
 
+    constructor(props) {
+        super(props);
+        throwIf(
+            this.model.switcherPosition != 'bottom',
+            'Mobile TabContainer only supports a bottom tab switcher at this time.'
+        );
+    }
+
     render() {
         const {model} = this,
-            {activeTabIndex} = model;
+            {activeTab} = model,
+            tabs = model.tabs.filter(it => !it.excludeFromSwitcher);
 
         return tabbar({
-            index: activeTabIndex,
-            renderTabs: () => model.tabs.map(tabModel => this.renderTab(tabModel)),
-            onPreChange: (event) => model.setActiveTabIndex(event.index)
+            index: activeTab ? tabs.indexOf(activeTab) : 0,
+            renderTabs: () => tabs.map(tabModel => this.renderTab(tabModel)),
+            onPreChange: (e) => model.activateTab(tabs[e.index].id)
         });
     }
 
     renderTab(tabModel) {
-        const {id, label, icon} = tabModel;
+        const {id, title, icon} = tabModel;
 
         return {
             content: tab({key: id, model: tabModel}),
@@ -38,7 +53,7 @@ export class TabContainer extends Component {
                 className: 'xh-tab',
                 items: [
                     div({className: 'xh-tab-icon', item: icon, omit: !icon}),
-                    div({className: 'xh-tab-label', item: label})
+                    div({className: 'xh-tab-label', item: title})
                 ]
             })
         };